requirements.txt
caldav
playwright
icalendar
pytzpython-dotenv
flask
requests